Add functions to submit a request immediately or to the throttled queue
This is what all the refactoring was getting towards. Now callbacks can
generate more requests that will be timed. If they SendRequest() then
the throttle controls the total number of requests sent; if they
SendRequestImmediately() then the secondary requests won't be counted
and the throttle controls the number of initial requests.